requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
docx2pdf==0.1.8
fire==0.5.0
Flask==3.0.0
fonttools==4.43.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
numpy==1.26.0
opencv-python==4.8.1.78
pdf2docx==0.5.6
Pillow==8.3.2
PyMuPDF==1.23.4
PyMuPDFb==1.23.3
python-docx==0.8.11
six==1.16.0
termcolor==2.3.0
python32==3.0.6
tqdm==4.66.1
Werkzeug==3.0.0
